Changed the login logic in the controller
